templates.js refactor
Some of the changes:

magicUpdate should be much faster now. I split apart upadating status for typing and for dragging since there was a 
lot of redundancy and calls to $(".placedfield") which was slowing down ie.

also, I cache the calls to .placedfield when you focus on a field. That way, all typing is super fast.

and lots of cleanup and removal of old code