add ghostscript wrapper to help guard against gs vulnerabilities
Ths wrapper won't be used by default, you need to install it on your system to use it, and you need to point the config at it.
There are instructions in /bin/gswrapper/instructions.txt.
They're a bit rubbish at the moment, but Francisco is going to write some better ones.