use cryptocipher instead of AES
cryptocipher doesn't support CFB so we need to align data to 16 bytes
because CBC requires that. this can be error prone for some type of binary 
data, but since we use it only for flash messages, there will be no problem.
however, I guess using crypto-api and cfb mode instead is the ultimate goal.