do not manually add https:// to links in emails
this was not a problem. the problem was that reminders are instantianted
from action scheduler and thus they do not use happstack hostpart, but
the one that is defined in kontrakcja.conf. but since the static hostparts
on production and staging were defined without https:// (as skrivapa.se and
staging.skrivapa.se appropriately) this is why generated links were wrong.