initial sign up
todo:
- displaying signup page as a modal
- displaying number of remaining days of free trial (didn't do that because it
  kinda breaks layout, this needs to be discussed)
- decreasing number of free signatures when users sign
- choosing priceplan where there is no free signatures left or free trial has ended

this also breaks signing from sign view (because of the fact that due
to storing modals/flash messages in cookies we are limited to 4kB and
it just seems that adding <input>s for choosing account type exceeds
that. that also includes displaying 'normal' signup as a modal (it
contains terms of agreement). I have to fix that.