Add user stats
The only stat recorded now is when a user signs TOS, which is considered the
first time the user takes an action (as opposed to fake user accounts which
we create when they receive a doc).

Migration needs to happen /adminonly/runstatsonallusers

There is a hack in that existing Service users (who have a termsofservice
time of 0) are counted as yesterday. It's more accurate than 1970.

New users will be added with the correct time. We just lost the
existing users' times.