Move doc upload to backbone
I did this primarily to fix the Entity too large message. It turned
out that I had to refactor a lot of stuff into backbone.

Now it uses backbone to manage the process instead of making round
trips to the server. It uses the API to create the document.

This made the templates simpler. I also removed some javascript from
the templates. It could be that some Haskell code for generating
those templates could be removed as well.

Usability improvements: Faster response on clicks; flashmessage for
unknown file type (non-pdf); flashmessage for too large a file.