Fix download JSON bug.
WHen you did an upload, on Firefox and sometimes on Chome, it would open
a save-as dialog and let you download a JSON file. I changed the content-type
on the API calls to text/plain so that doesn't happen. It was application/json,
which is technically the correct mime type for JSON api calls.

text/plain will do