Give UserInfo company name and company number
This is in the case where they don't want to use an entire
Company just for one guy who happens to work for himself.

It adds a migration to users table to add the two fields.
It also adds an instance of HasSomeCompanyInfo to User, UserInfo,
and (User, Maybe Company), which uses the one from the company
if it is Just, and the one from the User if Nothing.

Also adds form fields in the User Account and User administration
forms, which don't display when a company exists for that user.