Replace the promote scrive view in pos sign view with this modal on the design view.
The content is similar, but the design is nicer and it's easier
to get going.

Update to Jac's design and to use swedish and english words, not just english ones.

Don't use internals of the button.

Don't use internals of confirmation to hide the footer and don't use .popup(), use new.

Clean up previous modals that are no longer displayed in design view.