fix ordering in archive/document history/a few other places
previous library had only left fold, which was fetching stuff in reversed order,
so the common workaround was to sort with DESC to cancel the reversion. now it
was replaced with right fold in most places, but I forgot to remove DESC from
a few queries, hence some stuff ended up in reversed order.