Change replyTo from list of addresses to Maybe Address
Following a review by Mariusz we decided that the use-case for multiple
reply-to values is rare. Additionally the semantics of multiple reply-to is
unclear.
Finally to do things in proper relational DB way we would need to move
the list of addresses to a new table.

Current Development #55
See: https://trello.com/c/apxAdIPJ