Add function to catch error response for send SMS pin modal
Did not have any function to catch potential error response.
This could happen, for example, if the authentication method was changed
after the signatory opened the page.

The function used did not have anything specific to the view so made it
window.Blah instead so it is easy to call it.