Add previous field value to update field query and modify evidence log
Some changes to evidence log for checkbox changes, signature field and
text fields (including custom).

* For checkbox:
{name of person} ({email-address, phone number or personal number})
{unchecked | checked} the checkbox "{name of checkbox}". The checkbox's
placement in the Main Document:
{placements}

* For signature fields:
{name of person} ({email-address, phone number or personal number})
signed in the signature box "{name of box}". The box's placement in the
Main Document:
{placements}

* For text fields:
{name of person} ({email-address, phone number or personal number})
changed the text in the text field "{name of field}" from “{previous
value}” to “{new value}”. The field's placement in the Main Document:
{placements}

Also use word "blank" (no quotes) when the previous or new value is an
empty string.

See Trello card for more details.
Current Development #119
https://trello.com/c/zmoS77TO