Field placements cleanup as part of 2px border
Replace magic number font sizes with global variable at top of file for
visibility.

Move margins and border values from CSS to JS.
This may seem counter to what we want but in this case we need to
calculate field placement using these values and having them inlined in
generated HTML is actually better as we don't duplicate values that then
become out-of-sync.

Created new global vars in fieldplacementview.js that can be used to
change these.

Related to task in Current Development:
https://trello.com/c/NG1Kv9j0/127-make-border-2px-e-v1-t1
and also QA bug:
https://trello.com/c/dYtpRPK1/676-field-placement-in-design-view-is-off-by-a-few-pixels