Show save action feedback in designview
Add unit tests for the 'saved' document flag in JSON to ensure proper
behaviour of update call.
Once a document is 'saved' it should not be able to set it back to not
saved.
If doc is created through OAuth it should be saved when created,
otherwise it shouldn't be.

https://trello.com/c/vawQlhN5/138-show-feedback-on-all-save-action-in-design-view-e-v1-t1