API v2: Proper API Parameter functions in Parameters.hs
Using the magic of GADTs we can have different contructors for a API
Parameter data type and pattern match on it in a single function.

We use this to pass the proper data type to the same function and make
getting and parsing of different API parameters painless and beautiful.

* Change Parameters.hs following review by Mariusz (MR)
- Add export list instead of exporting everything in the module
- Rename 'getName' function and make it global within the module
- Change 'data ParameterOption a' to have three constructors and that is
  what the nested 'Maybe a' does anyway but is confusing
  -> update usage and functions to match this change
- Add 'ApiV2ParameterText'
- Also refactored the code a bit.