API v2: Implement Document "setfile" call
Does not change the document title like we do in API v1.
For frontend migrating to API v2, to keep the current behaviour, we will
need to do a separate Update call.

Reasoning:
This is unexpected behaviour from an API call to change the main file,
it affects a part of the document that is not involved in the API call.