API v2: Add a helper Guard to check for DocumentStatus
We are currently usig (Document -> Bool) from DocInfo to check document
status but this leads to inconsistent usage and error messages.
Also the code is not 100% clear.

With `guardDocumentStatus` it is clear from reading the API call what we
want to check and we don't need to worry about proper error text.