API v2: Replace processPDFParameter & ApiV2ParameterInput with ApiV2ParameterFile
I introduced `ApiV2ParameterInput` which could then be used with
`processPDFParameter` from CallsUtils.hs to get a FileId for a PDF file
given as part of the request.
This was strange as it needed the parameter name for proper error
messages, it needed extra case statements to handle, and was just odd.

Now it is where it truly belongs: in Parameters.hs

So I've replaced `ApiV2ParameterInput` with `ApiV2ParameterFile` which
gives you back a full-fledged File
Amazing :)

Thank you to MR for review :)