API v2: Rename functions and refactor working of Doc.API.V2.Parameters
Before we had `apiV2Parameter` and `apiV2Parameter'`, which is not very
clear in terms of naming.
Also `apiV2Parameter'` could have unexpected behaviour if the optional
parameter didn't have a default value as it would fail, even though you
thought it would be optional.

Now we have three functions with clear names and types:
  - apiV2ParameterOptional   :: ApiV2Parameter a -> m (Maybe a)
  - apiV2ParameterObligatory :: ApiV2Parameter a -> m a
  - apiV2ParameterDefault    :: a -> ApiV2Parameter a -> m a