Add document list CSV download to Archive.Control using ApiV2 parameters
We need document list CSV for our frontend.
We don't want this to be a proper API call as the CSV generation is
messy and we don't want to change the format.

A good place for this is along its "cousin": the ZIP archive download.

Uses List parameters from API v2 but returns same CSV format as V1 (and
depends on that code)

https://trello.com/c/l8hugeuU/371-move-frontend-to-apiv2