Change URL endpoints for branding resources so that Context is not needed
Make all branding resources use a URL endpoint so as to not require
information from Context, instead the URL will have the
BrandingDomainID, UserID, etc.
This is needed to forward these assets via CDN.
As a result branding assets will be "insecure" in that searching the URL
space you may find some logos... this is OK.
Change usage in String Templates.

TODO
* There is some JS usage that needs to be fixed
* Remove cdnurl from generation of css from less as we don't need it
  anymore

https://trello.com/c/u7yXezRx/410-branding-images-and-css-in-cdn