Add support for no-configuration errors and warnings
This patch is part of a series of patches to permit the configurations
for LiveDocx, CGI Group, Recurly, Mixpanel, HubSpot, Google Analytics
and Salesforce to be left out for test and development.

This patch adds two commands to promote consistent errors and warnings,
KontraError.noConfigurationError and
KontraError.noConfigurationWarning.

The module KontraError was chosen out of convenience for lack of a
better reason. Please feel free to move the commands somewhere else.