Make configuration optional for LiveDocx
This patch is part of a series of patches to permit the configurations
for LiveDocx, CGI Group, Recurly, Mixpanel, HubSpot, Google Analytics
and Salesforce to be left out for test and development.

This patch lifts the fields for LiveDocx in AppConf and Context to
Maybe as well as extends
Doc.API.V1.Calls.apiCallV1CreateFromFile,
Doc.API.V1.Calls.apiCallV1ChangeMainFile and
Doc.API.V2.Parameters.apiV2ParameterOptional
with case-of and noConfigurationError. The errors are caught by the API.