Make configuration optional for CGI Group
This patch is part of a series of patches to permit the configurations
for LiveDocx, CGI Group, Recurly, Mixpanel, HubSpot, Google Analytics
and Salesforce to be left out for test and development.

This patch lifts the fields for CGI Group in AppConf and Context to
Maybe as well as extends
Doc.Signing.Consumer.documentSigning,
EID.CGI.GRP.Control.handleAuthRequest,
EID.CGI.GRP.handleSignRequest and
EID.CGI.GRP.checkCGIAuthStatus,
with case-of, noConfigurationError and noConfigurationWarning. The
warning is logged by documentSigning, not to interfere with cron, while
the errors are caught by the API.