Make configuration optional for Recurly
This patch is part of a series of patches to permit the configurations
for LiveDocx, CGI Group, Recurly, Mixpanel, HubSpot, Google Analytics
and Salesforce to be left out for test and development.

This patch lifts the fields for Recurly in AppConf and Context to Maybe
as well as extends
Cron.main,
Administration.AdministrationControl.companyPaymentsJSON,
Payments.Control.handleSyncNewSubscriptionWithRecurly,
Payments.Control.handleChangePlan,
Payments.Control.postBackCache,
Payments.Control.handlePricePageUserPlanRecurlyJSON,
Payments.Control.handlePricePageUserPlanNoneJSON,
Payments.Control.handlePricePageUserJSON,
Payments.Control.handlePricePageNoUserJSON and
Payments.Control.handleSyncNewSubscriptionWithRecurlyOutside,
with case-of, noConfigurationError and noConfigurationWarning. The
warning is logged by main, not to interfere with cron, while the errors
are caught by the API.