Make configuration optional for Mixpanel
This patch is part of a series of patches to permit the configurations
for LiveDocx, CGI Group, Recurly, Mixpanel, HubSpot, Google Analytics
and Salesforce to be left out for test and development.

This patch lifts the fields for Mixpanel in AppConf and Context to
Maybe as well as extends
Cron.main,
with case-of and noConfigurationWarning. The warning is logged by main,
not to interfere with cron.