Make configuration optional for Salesforce
This patch is part of a series of patches to permit the configurations
for LiveDocx, CGI Group, Recurly, Mixpanel, HubSpot, Google Analytics
and Salesforce to be left out for test and development.

This patch lifts the fields for Salesforces in AppConf and Context to
Maybe as well as extends
Doc.API.Callback.Execute.execute,
Salesforce.Control.handleSalesforceIntegration,
Salesforce.Control.getSalesforceKeys and
User.API.apiCallTestSalesforceIntegration,
with case-of, noConfigurationError and noConfigurationWarning. The
warning is logged by execute, not to interfere with cron, while the
errors are caught by the API.

The functions have also been reformatted as part of untangling the
logic for Salesforce. Moreover, HasSalesforceConf has been removed.