Remove optional parts of vagrant/kontrakcja.conf
This patch is part of a series of patches to permit the configurations
for LiveDocx, CGI Group, Recurly, Mixpanel, HubSpot, Google Analytics
and Salesforce to be left out for test and development.

This patch removes those parts of vagrant/kontrakcja.conf that can be
left out as to provide a minimal working setup.