Rename fields in log objects
Make this so that log paths "type check". For all log objects all paths must
be used with same type of JSON value. If they don't, Kibana tries to parse
Object as an Int or String as Array or ... and then it breaks.
Since we are not checking this (and we cannot for JSONs coming from external
libraries), this will break again soon.

case 2653