Improve post-document-closed actions (#533)
* Improve post-document-closed actions

When GT signing of the document failed, the job was not re-scheduled.
To fix that, there was a cron job, that searched for such documents. The query
took a long time (~20secs)
Now the document GT signing is re-scheduled on failure. The cron job was removed
because it's not necessary anymore.

case 2645

* Small adjustments from review