Refactor and expand Signatory Fields documentation
Stop using a common `SignatoryField` definition that maps badly to the
various "concrete" definitions.

Instead move the common parts to `SignatoryFieldStandardPlacements`,
which in turn use `SignatoryFieldPlacementTip` and
`SignatoryFieldPlacementAnchor`.

Added more comprehensive descriptions as part of the rewrite, which
should make it easier to understand Signatory Fields.

First part for Case 3271.