Rewrite ScreenBlockingDialog to React
LoadingDialog uses ScreenBlockingDialog behind the scenes so it
automatically uses the new React components.

As an added bonus, this commit fixes an obscure bug in signatory
attachments view in sign view, where an error message wouldn't
display because it was trying to use React button as if it was a
Backbone button.

Addresses case 2995