API -> Link Delivery and Compose email button on Authors view (#757)
* Change API delivery to Link delivery in Frontend

* Update sprite-participation-icons to version without API text
* Update localization

No functionality change, and we don't change any Document JSON.
This is purely a UI change, it is easier to understand "link delivery"
and it allows us to complete CORE-263.

* Add "Compose email" button to Author view for Link delivery when email set

CORE-263

RBS needs to send invitation mail through their own mail server. The
easiest way to enable this for them is to let them send mail manually.
But copy-pasting the partial link provided in author view using delivery
method:API is a bit cumbersome. So instead we could add a "self-service
e-mail" delivery method. This gives the same result as API delivery
method except we give a mailto link containing some text and the
complete URL to the signing page for the signatory.

* Make texts into proper HTML for "compose email" button (Link delivery)

CORE-263

* Fix linting issues