Shake script improvements (#852)
* Formatting.

* Shake: allow selecting which tests to run.

So you can now do, for example:

    ./shake.sh --new-build test-server -p AccountInfo -p
    AdministrationControl

I picked '-p' instead of '-t' for consistency with Tasty (we'll want
to migrate the test suite to it in the future anyway).

* Shake: add the component dependency graph to CabalFile.

But don't make use of it yet.

* Drop Cabal < 2.0 support.

GHC 8.2+ ships with Cabal >= 2.0.

* Refactoring: get rid of CabalComponent{Name,Type}.

Use types from the Cabal library directly.

* Shake: respect dependencies betweeen components.

So e.g. if you make a change to 'lib:kontrakcja' and then try to run
'./shake.sh run-server-tests', the Shake script will now correctly
rebuild 'lib:kontrakcja' before running the test suite.

* sort_imports: Don't treat Prelude as a local module.

* Shake: Cabal 2.0 and GHC < 8.4 compat.

* Shake: add rules for exe components.

Also change the component dependency rule to depend on sources instead
of component names -- we only need to run 'cabal new-build foo' if
'foo' was requested, and its dependency 'bar' was changed.