CORE-948 Shareable links feature flag
Sets up a frontend (only) enforced feature flag.

The feature flag removes the icons relating to shareable links from the
templates page if it has been disabled for the user.

Note that since it is only frontend-enforced, a determined user could
circumvent it by sending appropriate HTTP requests to the back end.