Fix issue front-end tests to run on newest node
An warning is caused by the deprecation of the Buffer() constructor.
This warning causes the tests to fail.

A guide to fixing the issue is given here:

https://nodejs.org/en/docs/guides/buffer-constructor-deprecation/

I've followed these instructions and it appears to have fixed the issue without
breaking the tests for older versions of node. (Tested with node v8.11.4)