CORE-1016 - Secure QR Codes for temp login tokens
This commit implements temporary login_tokens by using the existing
infrastructure for `MagicHash`s and the generation of QR code PNGs.

These tokens can be set to have a range of expiry times with the constraint that
1 <= n <= 30. Currently, the 30 minute maximum validity time is hard-coded into
the `apiCallGetTokenForPersonalCredentials` handler definition where clause.

This could be made configurable via kontrakcja.conf in the future. A maximum
upper bound on validity time is important to ensure the security of the tokens.

Currently, there is no limit to the number of tokens that can be active
simultaneously for a given user. Cancellation of tokens is thereby handled by
simply allowing it to expire.

A new database table `temporary_login_tokens` is created to store the tokens.

`apiCallGetUserPersonalToken` has also been extended so that it can accept a
login_token in lieu of (but not as well as) a username/password combination.

If User A has "Update" access privileges for User B, User A can generate a
login_token for User B. Given that "Update" privileges allow A to turn off 2FA
for User B anyway, `login_token`s do not require the provision of a 2FA token
when used with `apiCallGetUserPersonalToken` since this would provide no extra
security.

This commit also contains a small amount of refactoring.

+ `apiAccessControl` has been moved to a util module (API.V2.Utils)
+ `noPrvErr` has been eliminated and hard coded into the def of
  `apiAccessControl` since it is always (and only) used in conjuction with this
  function.

A cron job has also been added which regularly purges the expired login tokens.

Tests have also been written to ensure the correct function of most of the new
functionality of login_tokens.