Adjust the order of migrations for staging
The DB schemas of master and staging has diverged. Explanation from slack:

Ok, so when I merge staging into master, I'm going to adjust the order of
migrations so that it's correct for master, incorrect for staging. Then I'm
going to merge master into staging. so staging has incorrect order of
migrations. Then I'm going to commit a fix (this is the fix) to staging for the
migration order. That way both master and staging is migrated correctly and TC
won't complain. Then, on the next staging<->master sync, staging is merged into
master first so that the order of migrations for master is changed (and thus
synchronized between 3 branches), but it won't matter then because the db has
already been migrated. It will break migrations for people who had master before
today's staging->master merge and didn't pull until the next merge
staging->master happens, but that's somewhat improbable and results in a local
breakage only, so it's not the end of the world.