CORE-1354 CORE-978 Bare minimum UserGroup API
A design decision was made to have REST-like endpoints which take JSON via a
formData field rather than as a raw request body. Only GET and POST verbs are
used too, actions like "delete" are signified in the URL.

Other than these deviations from convention, the endpoints are roughly RESTful.

This commit is a stripped down version of a larger piece of work that allows us
to get the usergroup GET endpoint reviewed and merged more quickly. The rest of
the endpoints will be merged in a later commit.

This commit contains:

+ A set of `UnjsonDef`s that specify the JSON structure for `UserGroup` objects
  (and their subtree objects).
+ Some refactoring in `UserGroup.Internal` to allow for a `defaultUserGroup` and
  a `detaultChildUserGroup` object. The subtrees, such as `defaultUserGroupUI`.
  have also been split out from `defaultUserGroup` and externally exposed.
    + This makes this commit a solution to CORE-1263.
+ Tests check that access control performs as expected for the GET endpoint
+ The way in which the Access Control requirement objects are constructed is
  improved via existence checks. New helpers perform this task:
    + `makeAPIUserAccessPolicyReq`
    + `makeAPIUserGroupAccessPolicyReq`

+ The above helpers return the thing being existence checked.