[CORE-2069] SEBankID auth-to-view via eID Hub (#1550)
* Frontend: SEBankID identify via EID Hub

* Backend: SEBankID identify via EID Hub

* Make SEBankID configurable per user group

* Fix SEBankID on iOS in case we land in a different browser

On iOS there is a possible scenario where we land in a different browser
after concluding the transaction with EID Hub (because the BankID app on
iOS is weird); see https://scriveab.atlassian.net/browse/AUTH-201. We
can handle this by including the session ID in the redirect URL, and
restoring the session from that.