[CORE-2569] Fix communication with EID Hub (in particular large requests) (#1839)
* Fix curl parsing of http response code

Sometimes a single curl call results in more than one http response
code, in particular if the request is large there is an intermediate
HTTP 100. This breaks our hack of parsing the headers returned with
`-i`. The fix is to use `-w "\n%{http_code}"` to make curl attach the
_last_ http response code after the response content.

* Don't log content of eid hub transaction

With Verimi QES the content contains the pdf being signed, which is both
big and very sensitive information.